back/requirements.txt
aiohttp==3.8.3
fastapi==0.91.0
uvicorn==0.20.0
redis>=4.2.0rc1
python-socketio
websockets
